[11.x] Fix the chunk method to an integer type in the splitIn method (#51733)

+++ src/Illuminate/Collections/Collection.php
@@ -1254,7 +1254,7 @@ class Collection implements ArrayAccess, CanBeEscapedWhenCastToString, Enumerabl
      */
     public function splitIn($numberOfGroups)
     {
-        return $this->chunk(ceil($this->count() / $numberOfGroups));
+        return $this->chunk((int) ceil($this->count() / $numberOfGroups));
     }
 
     /**
